+++ RollbarFramework/RollbarFramework.h
@@ -8,16 +8,49 @@ FOUNDATION_EXPORT const unsigned char RollbarFrameworkVersionString[];
 
 #import <KSCrash/KSCrash.h>
 
+// Notifier API
 #import <RollbarFramework/Rollbar.h>
 #import <RollbarFramework/RollbarNotifier.h>
 #import <RollbarFramework/RollbarConfiguration.h>
-#import <RollbarFramework/RollbarThread.h>
-#import <RollbarFramework/RollbarFileReader.h>
 #import <RollbarFramework/RollbarReachability.h>
 #import <RollbarFramework/RollbarLogger.h>
-#import <RollbarFramework/RollbarLevel.h>
 #import <RollbarFramework/RollbarKSCrashInstallation.h>
 #import <RollbarFramework/RollbarKSCrashReportSink.h>
-#import <RollbarFramework/NSJSONSerialization+Rollbar.h>
 #import <RollbarFramework/RollbarTelemetry.h>
 #import <RollbarFramework/RollbarTelemetryType.h>
+
+// Deploys API
+#import <RollbarFramework/RollbarDeploysProtocol.h>
+#import <RollbarFramework/RollbarDeploysManager.h>
+
+#import <RollbarFramework/RollbarJSONFriendlyProtocol.h>
+#import <RollbarFramework/RollbarJSONFriendlyObject.h>
+#import <RollbarFramework/DeployApiCallResult.h>
+#import <RollbarFramework/Deployment.h>
+#import <RollbarFramework/DeploymentDetails.h>
+
+// DTOs:
+//======
+
+// Configuration
+#import <RollbarFramework/DataTransferObject.h>
+#import <RollbarFramework/DataTransferObject+CustomData.h>
+#import <RollbarFramework/RollbarLevel.h>
+#import <RollbarFramework/CaptureIpType.h>
+#import <RollbarFramework/RollbarConfig.h>
+#import <RollbarFramework/RollbarDestination.h>
+#import <RollbarFramework/RollbarDeveloperOptions.h>
+#import <RollbarFramework/RollbarProxy.h>
+#import <RollbarFramework/RollbarScrubbingOptions.h>
+#import <RollbarFramework/RollbarServer.h>
+#import <RollbarFramework/RollbarPerson.h>
+#import <RollbarFramework/RollbarModule.h>
+#import <RollbarFramework/RollbarTelemetryOptions.h>
+#import <RollbarFramework/RollbarLoggingOptions.h>
+// Payload
+// ...
+// Deploys
+// ...
+
+
+